@@ -16,11 +16,12 @@
     "@dbetka/wdk": "^0.1.5",
     "@egjs/flicking-plugins": "^4.2.2",
     "@egjs/vue3-flicking": "^4.4.1",
+    "@vuelidate/core": "^2.0.0-alpha.32",
+    "@vuelidate/validators": "^2.0.0-alpha.25",
     "dayjs": "^1.10.7",
     "js-cookie": "^2.2.1",
     "js-search": "^2.0.0",
     "ol": "^6.9.0",
-    "vee-validate": "^4.5.5",
     "vue": "^3.2.21",
     "vue-ellipse-progress": "^2.0.0",
     "vue-i18n": "^9.2.0-beta.17",
@@ -41,7 +42,7 @@
     "sass-loader": "^12.1.0",
     "style-loader": "^3.2.1",
     "ts-loader": "^9.2.6",
-    "typescript": "^4.4.3",
+    "typescript": "^4.5.2",
     "url-loader": "^4.1.1",
     "vue-loader": "^16.8.3",
     "webpack": "^5.49.0",
