@@ -6,7 +6,7 @@
     <o-form :on-submit="collectPoint">
       <m-field-text
         :label="$t('form.field.pointId')"
-        :rules="rules.pointId"
+        :rules="validationRules.pointId"
         v-model="collectedPointId"
         :disabled="blockForm"
       />
@@ -21,50 +21,61 @@
 <script>
 import TPage from 'templates/page';
 import { ROUTES } from 'config/routes-config';
-import { mixins } from 'mixins/base';
 import OForm from 'organisms/form';
 import MFieldText from 'molecules/field/text';
 import AButtonSubmit from 'atoms/button/submit';
 import { translator } from 'src/dictionary';
-import { materialIcons } from '@dbetka/vue-material-icons';
-
-const ICONS = materialIcons.names;
+import { useIcons } from '@dbetka/vue-material-icons';
+import { ref } from 'vue';
+import { useForm } from 'plugins/form';
+import { useStore } from 'vuex';
+import { useRouter } from 'vue-router';
 
 export default {
   name: 'p-collect-point',
-  mixins: [mixins.form, mixins.validation],
   components: {
     TPage,
     AButtonSubmit,
     MFieldText,
     OForm,
   },
-  data: () => ({
-    collectedPointId: '',
-    blockForm: false,
-    isSending: false,
-  }),
-  methods: {
-    onCollectPoint () {
-      this.$store.dispatch('popup/openTemporary', {
+  setup () {
+    const collectedPointId = ref('');
+
+    const form = useForm();
+    const { isSending, blockForm, onErrorOccurs } = form;
+
+    const store = useStore();
+    const router = useRouter();
+    const icons = useIcons();
+
+    function onCollectPoint () {
+      store.dispatch('popup/openTemporary', {
         messages: [
           translator.t('communicate.collectPoint.congratulation'),
           translator.t('communicate.collectPoint.youCollectedPoint'),
         ],
-        icon: ICONS.check_circle_outline,
+        icon: icons.names.check_circle_outline,
       })
         .then(() => {
-          this.isSending = false;
-          this.blockForm = false;
-          this.$router.push(ROUTES.map.path);
+          isSending.value = false;
+          blockForm.value = false;
+          router.push(ROUTES.map.path);
         });
-    },
-    collectPoint () {
-      const pointId = this.collectedPointId;
-      this.$store.dispatch('event/collectPoint', pointId)
-        .then(this.onCollectPoint)
-        .catch(this.onErrorOccurs);
-    },
+    }
+
+    function collectPoint () {
+      const pointId = collectedPointId.value;
+      store.dispatch('event/collectPoint', pointId)
+        .then(onCollectPoint)
+        .catch(onErrorOccurs);
+    }
+
+    return {
+      collectedPointId,
+      collectPoint,
+      ...form,
+    };
   },
 };
 </script>
