@@ -21,13 +21,13 @@ import TPage from 'templates/page';
 import OForm from 'organisms/form';
 import MInput from 'molecules/input';
 import AButtonSubmit from 'atoms/button/submit';
-import { mixins } from 'mixins/base';
 import OPopupEventConfirmation from 'organisms/popup/event-confirmation';
 import { eventsListMock } from 'organisms/events-list-mock';
+import { ref } from 'vue';
+import { useForm } from 'plugins/form';
 
 export default {
   name: 'p-join-event',
-  mixins: [mixins.form],
   components: {
     OPopupEventConfirmation,
     TPage,
@@ -35,16 +35,23 @@ export default {
     MInput,
     AButtonSubmit,
   },
-  data: () => ({
-    eventCode: '',
-    blockForm: false,
-    isSending: false,
-    event: eventsListMock[0],
-  }),
-  methods: {
-    toggleDetails () {
-      this.$refs.popupScore && this.$refs.popupScore.toggle();
-    },
+  setup () {
+    const eventCode = ref('');
+    const event = ref(eventsListMock[0]);
+    const popupScore = ref(null);
+    const form = useForm();
+
+    function toggleDetails () {
+      popupScore.value && popupScore.value.toggle();
+    }
+
+    return {
+      ...form,
+      eventCode,
+      event,
+      popupScore,
+      toggleDetails,
+    };
   },
 };
 </script>
