@@ -21,16 +21,15 @@
 </template>
 
 <script>
-import { mixins } from 'mixins/base';
+import { modelValueMixin, useModelValue } from 'plugins/v-model';
+import { computed, ref, toRefs } from 'vue';
+import { useIcons } from '@dbetka/vue-material-icons';
 
 export default {
   name: 'a-checkbox',
-  mixins: [mixins.vModelCheckbox],
-  data: () => ({
-    hasFocus: false,
-  }),
   props: {
     id: {
+      type: [String, Number],
       default: '',
     },
     isDisabled: {
@@ -38,19 +37,31 @@ export default {
       default: false,
     },
   },
-  computed: {
-    labelClass () {
-      return {
-        'f-flex': true,
-        'f-disabled': this.isDisabled,
-        'f-text-primary': this.hasFocus,
-      };
-    },
-    iconName () {
-      const checkedIcon = this.$icons.names.check_box;
-      const uncheckedIcon = this.$icons.names.check_box_outline_blank;
-      return this.checked ? checkedIcon : uncheckedIcon;
-    },
+  mixins: [modelValueMixin],
+  setup (props, context) {
+    const { vModel } = useModelValue(props, context);
+    const { isDisabled } = toRefs(props);
+    const hasFocus = ref(false);
+
+    const labelClass = computed(() => ({
+      'f-flex': true,
+      'f-disabled': isDisabled.value,
+      'f-text-primary': hasFocus.value,
+    }));
+
+    const iconName = computed(() => {
+      const iconsNames = useIcons().names;
+      const checkedIcon = iconsNames.check_box;
+      const uncheckedIcon = iconsNames.check_box_outline_blank;
+      return vModel.value ? checkedIcon : uncheckedIcon;
+    });
+
+    return {
+      vModel,
+      hasFocus,
+      labelClass,
+      iconName,
+    };
   },
 };
 </script>
