@@ -1,30 +1,29 @@
 <template>
-  <!--  <validation-observer>-->
-  <!--    <validation-provider-->
-  <!--      :name="label.toLowerCase()"-->
-  <!--      :rules="rules.email"-->
-  <!--      v-slot="{ errors }"-->
-  <!--    >-->
   <m-input
     type="email"
     v-model.trim="vModel"
     :disabled="disabled"
     :placeholder="label"
+    :error="isError"
+    :assist="errorMessage || assist"
   />
-<!--  :error="errors.length > 0"-->
-<!--  :assist="errors[0] || assist"-->
-<!--    </validation-provider>-->
-<!--  </validation-observer>-->
 </template>
+
+<!-- USAGE EXAMPLE
+  <m-field-email
+    v-model="values.user"
+    :disabled="blockForm"
+  />
+-->
 
 <script>
 import MInput from 'molecules/input';
-import { mixins } from 'mixins/base';
 import { translator } from 'src/dictionary';
+import { validationRules } from 'plugins/validation/rules';
+import { fieldValidationMixin, useFieldValidation } from 'plugins/validation/field';
 
 export default {
   name: 'm-field-email',
-  mixins: [mixins.vModel, mixins.validation],
   components: { MInput },
   props: {
     disabled: Boolean,
@@ -37,5 +36,11 @@ export default {
       default: '',
     },
   },
+  mixins: [fieldValidationMixin],
+  setup: (props, context) => ({
+    ...useFieldValidation(props, context, {
+      defaultRules: validationRules.email,
+    }),
+  }),
 };
 </script>
