@@ -1,17 +1,20 @@
 <template>
-  <!--  <validation-observer tag="div" v-slot="{ handleSubmit }">-->
   <form
     class="f-flex f-flex-col f-pb-1 f-text-left"
+    @submit.prevent="handleSubmit()"
   >
-    <!--    @submit.prevent="handleSubmit(onSubmit)"-->
     <slot v-if="isSend === false"/>
     <slot name="form" v-if="isSend === false"/>
     <slot v-else name="response"/>
   </form>
-<!--  </validation-observer>-->
 </template>
 
 <script>
+import useVuelidate from '@vuelidate/core';
+import { store } from 'store';
+import { toRefs } from 'vue';
+import { translator } from 'dictionary';
+
 export default {
   name: 'o-form',
   props: {
@@ -23,6 +26,27 @@ export default {
       type: Boolean,
       default: false,
     },
+  },
+  setup (props) {
+    const v$ = useVuelidate();
+    const { onSubmit } = toRefs(props);
+
+    function handleSubmit () {
+      if (v$.value.$invalid) {
+        v$.value.$touch();
+        store.dispatch('snackbar/openTemporary', {
+          message: translator.t('error.correctForm'),
+          error: true,
+        });
+      } else {
+        onSubmit.value();
+      }
+    }
+
+    return {
+      handleSubmit,
+      v$,
+    };
   },
 };
 </script>
