@@ -1,4 +1,4 @@
-import { compareDate, splitObjectsListByTime } from 'utils/date';
+import { compareDate, displayDate, getDate, splitObjectsListByTime } from 'utils/date';
 
 const { isToday, isActual, isFuture, isPast } = compareDate;
 
@@ -21,8 +21,28 @@ const listUtils = {
   splitEventsByTimePeriods: (eventsList) => splitObjectsListByTime(eventsList, 'eventStartDate', 'eventEndDate'),
 };
 
+const conversions = {
+  convertEventToForm (oldData) {
+    const data = { ...oldData };
+    if (oldData.eventStartDate && oldData.eventEndDate) {
+      data.eventStartDate = getDate.fromTimestamp(oldData.eventStartDate);
+      data.eventEndDate = getDate.fromTimestamp(oldData.eventEndDate);
+    }
+    return data;
+  },
+  convertEventToSend (oldData) {
+    const data = { ...oldData };
+    if (oldData.eventStartDate && oldData.eventEndDate) {
+      data.eventStartDate = displayDate.inTimestamp(oldData.eventStartDate);
+      data.eventEndDate = displayDate.inTimestamp(oldData.eventEndDate);
+    }
+    return data;
+  },
+};
+
 export const eventUtils = {
   ...timeUtils,
   ...listUtils,
   ...stateUtils,
+  ...conversions,
 };
